apps/export/export_service/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
httpx>=0.24.0
PyJWT>=2.6.0
python-multipart>=0.0.6
weasyprint>=59.0
reportlab>=4.0.0
python-docx>=0.8.11
jinja2>=3.1.2
tenacity>=8.2.2
asyncio>=3.4.3
aiofiles>=23.1.0 